install test

+++ install.sh
@@ -18,7 +18,7 @@ if [ "$result" = true ]; then
   ! [ -x "$(command -v nvim)" ] && brew install nvim
   rm -rf ~/.local/share/nvim/site/pack/packer/start/packer.nvim
   git clone --depth 1 https://github.com/wbthomason/packer.nvim ~/.local/share/nvim/site/pack/packer/start/packer.nvim
-  rm -rf ~/.config/nvim.old
+  [ -d ~./config/nvim.old ] && rm -rf ~/.config/nvim.old
   [ -d ~/.config/nvim ] && mv -f ~/.config/nvim ~/.config/nvim.old 
   ln -s $INSTALL_DIR/nvim ~/.config/nvim
   mkdir -p ~/.config/nvim/undo
@@ -27,7 +27,7 @@ fi
 # zshrc
 prompt "Do you want to install my .zshrc?"
 if [ "$result" = true ]; then
-  [ -f ~/.zshrc.old ] rm ~/.zshrc
-  [ -f ~/.zshrc ] mv ~/.zshrc ~/.zshrc.old
+  [ -f ~/.zshrc.old ] && rm ~/.zshrc
+  [ -f ~/.zshrc ] && mv ~/.zshrc ~/.zshrc.old
   ln -s $INSTALL_DIR/.zshrc ~/.zshrc
 fi
